# perf: reduce loops in getValidOptions / separate to getValidReports