@hannojg 2025-01-16
I didn't had the time to record videos for all platforms - grateful for if the reviewer could help here - thanks! 🙏